+++ datapipe/store/filedir.py
@@ -1,5 +1,5 @@
 from abc import ABC
-from typing import IO, Any, Dict, List, Union, cast, Iterator
+from typing import IO, Any, Dict, List, Optional, Union, cast, Iterator
 from pathlib import Path
 
 import numpy as np
@@ -10,7 +10,7 @@ import json
 import fsspec
 import pandas as pd
 
-from sqlalchemy import Column, String
+from sqlalchemy import Column, Integer, String
 from PIL import Image
 
 from datapipe.types import DataDF, DataSchema, MetaSchema, IndexDF
@@ -108,7 +108,9 @@ class TableStoreFiledir(TableStore):
         adapter: ItemStoreFileAdapter,
         add_filepath_column: bool = False,
         primary_schema: DataSchema = None,
-        read_data: bool = True
+        read_data: bool = True,
+        readonly: bool = False,
+        disable_rm: Optional[bool] = True
     ):
         """
         При построении `TableStoreFiledir` есть два способа указать схему
@@ -138,23 +140,29 @@ class TableStoreFiledir(TableStore):
 
         read_data -- если False - при чтении не происходит парсинга содержимого
         файла
+
+        readonly -- если True, отключить запись файлов
+        disable_rm -- если True, отключить удаление файлов
         """
 
-        protocol, path = fsspec.core.split_protocol(filename_pattern)
+        self.protocol, path = fsspec.core.split_protocol(filename_pattern)
+        self.filesystem = fsspec.filesystem(self.protocol)
 
-        if protocol is None or protocol == 'file':
+        if self.protocol is None or self.protocol == 'file':
             self.filename_pattern = str(Path(path).resolve())
             filename_pattern_for_match = self.filename_pattern
-            self.protocol_str = "" if protocol is None else "file://"
+            self.protocol_str = "" if self.protocol is None else "file://"
         else:
             self.filename_pattern = str(filename_pattern)
             filename_pattern_for_match = path
-            self.protocol_str = f"{protocol}://"
+            self.protocol_str = f"{self.protocol}://"
 
-        if '*' in path:
-            self.readonly = True
-        else:
-            self.readonly = False
+        if '*' in path and readonly:
+            raise ValueError(
+                "When `disable_rm=False`, in filename_pattern shouldn't be any `*` characters."
+            )
+        self.readonly = readonly
+        self.disable_rm = disable_rm
 
         self.adapter = adapter
         self.add_filepath_column = add_filepath_column
@@ -164,15 +172,24 @@ class TableStoreFiledir(TableStore):
         self.filename_glob = _pattern_to_glob(self.filename_pattern)
         self.filename_match = _pattern_to_match(filename_pattern_for_match)
 
+        type_to_cls = {
+            String: str,
+            Integer: int
+        }
+
         if primary_schema is not None:
             assert sorted(self.attrnames) == sorted(i.name for i in primary_schema)
-
+            assert all([isinstance(column.type, (String, Integer)) for column in primary_schema])
             self.primary_schema = primary_schema
         else:
             self.primary_schema = [
                 Column(attrname, String(100), primary_key=True)
                 for attrname in self.attrnames
             ]
+        self.attrname_to_cls = {
+            column.name: type_to_cls[type(column.type)]
+            for column in self.primary_schema
+        }
 
     def get_primary_schema(self) -> DataSchema:
         return self.primary_schema
@@ -181,25 +198,34 @@ class TableStoreFiledir(TableStore):
         return []
 
     def delete_rows(self, idx: IndexDF) -> None:
-        # FIXME: Реализовать
-        # Do not delete old files for now
-        # Consider self.readonly as well
-        assert(not self.readonly)
+        assert(not self.disable_rm)
 
-        pass
+        for row_idx in idx.index:
+            _, path = fsspec.core.split_protocol(
+                self._filename_from_idxs_values(idx.loc[row_idx, self.attrnames])
+            )
+            self.filesystem.rm(path)
 
     def _filename_from_idxs_values(self, idxs_values: List[str]) -> str:
         return re.sub(r'\{([^/]+?)\}', Replacer(idxs_values), self.filename_pattern)
 
-    def _assert_key_values(self, filepath: str, idxs_values: List[str]):
+    def _idxs_values_from_filepath(self, filepath: str) -> Dict[str, Any]:
         _, filepath = fsspec.core.split_protocol(filepath)
         m = re.match(self.filename_match, filepath)
+        assert m is not None, f"Filepath {filepath} does not match the pattern {self.filename_match}"
 
+        data = {}
+        for attrname in self.attrnames:
+            data[attrname] = self.attrname_to_cls[attrname](m.group(attrname))
+
+        return data
+
+    def _assert_key_values(self, filepath: str, idxs_values: List[str]):
+        idx_data = self._idxs_values_from_filepath(filepath)
         idxs_values_np = np.array(idxs_values)
-        idxs_values_parsed_from_filepath = np.array([
-            m.group(attrname) for attrname in self.attrnames
-            if m is not None
-        ])
+        idxs_values_parsed_from_filepath = np.array(
+            [idx_data[attrname] for attrname in self.attrnames]
+        )
 
         assert (
             len(idxs_values_np) == len(idxs_values_parsed_from_filepath) and
@@ -207,7 +233,8 @@ class TableStoreFiledir(TableStore):
             np.all(idxs_values_np == idxs_values_parsed_from_filepath)
         ), (
             "Multiply indexes have complex contradictory values, so that it couldn't unambiguously name the files. "
-            "This is most likely due to imperfect separators between {id} keys in the scheme."
+            "This is most likely due to imperfect separators between {id} keys in the scheme.",
+            f"{idxs_values_np=} not equals {idxs_values_parsed_from_filepath=}"
         )
 
     def insert_rows(self, df: pd.DataFrame) -> None:
@@ -230,39 +257,43 @@ class TableStoreFiledir(TableStore):
                 self.adapter.dump(data, f)
 
     def read_rows(self, idx: IndexDF = None) -> DataDF:
-        assert(idx is not None)
+        def _iterate_files():
+            if idx is None:
+                for file_open in fsspec.open_files(self.filename_glob, f'r{self.adapter.mode}'):
+                    yield file_open
+            else:
+                for row_idx in idx.index:
+                    yield fsspec.open(
+                        self._filename_from_idxs_values(idx.loc[row_idx, self.attrnames]), f'r{self.adapter.mode}'
+                    )
+        df = []
+        for file_open in _iterate_files():
+            with file_open as f:
+                data = {}
 
-        def _gen():
-            for row_idx in idx.index:
-                with (file_open := fsspec.open(self._filename_from_idxs_values(idx.loc[row_idx, self.attrnames]),
-                                               f'r{self.adapter.mode}')) as f:
-                    data = {}
+                if self.read_data:
+                    data = self.adapter.load(f)
 
-                    if self.read_data:
-                        data = self.adapter.load(f)
+                    attrnames_in_data = [attrname for attrname in self.attrnames if attrname in data]
+                    assert len(attrnames_in_data) == 0, (
+                        f"Found repeated keys inside data that are already used (from scheme): "
+                        f"{attrnames_in_data}. "
+                        f"Remove these keys from data."
+                    )
 
-                        attrnames_in_data = [attrname for attrname in self.attrnames if attrname in data]
-                        assert len(attrnames_in_data) == 0, (
-                            f"Found repeated keys inside data that are already used (from scheme): "
-                            f"{attrnames_in_data}. "
-                            f"Remove these keys from data."
-                        )
+                idxs_values = self._idxs_values_from_filepath(file_open.path)
+                data.update(idxs_values)
 
-                    for attrname in self.attrnames:
-                        data[attrname] = idx.loc[row_idx, attrname]
+                if self.add_filepath_column:
+                    assert 'filepath' not in data, (
+                        "The key 'filepath' is already exists in data. "
+                        "Switch argument add_filepath_column to False or rename this key in input data."
+                    )
+                    data['filepath'] = f"{self.protocol_str}{file_open.path}"
 
-                    if self.add_filepath_column:
-                        assert 'filepath' not in data, (
-                            "The key 'filepath' is already exists in data. "
-                            "Switch argument add_filepath_column to False or rename this key in input data."
-                        )
-                        data['filepath'] = f"{self.protocol_str}{file_open.path}"
+                df.append(data)
 
-                    yield data
-
-        return pd.DataFrame.from_records(
-            _gen()
-        )
+        return pd.DataFrame(df)
 
     def read_rows_meta_pseudo_df(self, chunksize: int = 1000, run_config: RunConfig = None) -> Iterator[DataDF]:
         # FIXME реализовать чанкирование