+++ datapipe/store/table_store.py
@@ -27,7 +27,7 @@ class TableStore(ABC):
         raise NotImplementedError
 
     def update_rows(self, df: DataDF) -> None:
-        self.delete_rows(df.index)
+        self.delete_rows(data_to_index(df, self.primary_keys))
         self.insert_rows(df)
 
     def read_rows(self, idx: IndexDF = None) -> DataDF: